// vote_app_backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const VoteType = {
  FREE: 'FREE',
  PAID: 'PAID'
} as const

export type VoteType = (typeof VoteType)[keyof typeof VoteType]


export const EventStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  LIVE: 'LIVE',
  ENDED: 'ENDED',
  SUSPENDED: 'SUSPENDED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]


export const PlanType = {
  FREE: 'FREE',
  PACK: 'PACK',
  PREMIUM: 'PREMIUM'
} as const

export type PlanType = (typeof PlanType)[keyof typeof PlanType]


export const TransactionStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  DECLINED: 'DECLINED',
  REFUNDED: 'REFUNDED'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]


export const SubscriptionStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED'
} as const

export type SubscriptionStatus = (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus]


export const PayoutStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type PayoutStatus = (typeof PayoutStatus)[keyof typeof PayoutStatus]